refactor and add some comments
